requirements.txt
asgiref==3.11.1
Django==5.2.11
django-modeltranslation==0.19.19
pillow==12.1.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
